Fix some warnings

+++ engine/client/cl_cmds.c
@@ -347,7 +347,7 @@ void CL_LevelShot_f( void )
 	}
 
 	// missing levelshot or level newer than levelshot
-	if( ft2 == -1 || ft1 > ft2 )
+	if( ft2 == (unsigned long)-1 || ft1 > ft2 )
 		cls.scrshot_action = scrshot_plaque;	// build new frame for levelshot
 	else cls.scrshot_action = scrshot_inactive;	// disable - not needs
 }
